refactor: game and platform interfaces for improved type safety and organization
