+++ src/hooks/useGames.ts
@@ -2,20 +2,9 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 import ms from 'ms';
 import APIClient, { FetchResponse } from '../services/api-client';
 import { useGameQueryStore } from '../store';
-import { Platform } from './usePlatform';
+import { Game } from '../interfaces/Game';
 
 const apiClient = new APIClient<Game>('/games');
-
-export interface Game {
-  id: number;
-  slug: string;
-  name: string;
-  description_raw: string;
-  background_image: string;
-  parent_platforms: { platform: Platform }[];
-  metacritic: number;
-  rating_top: number;
-}
 
 const useGames = () => {
   const gameQuery = useGameQueryStore((s) => s.gameQuery);
