+++ src/hooks/useGenres.ts
@@ -2,20 +2,15 @@ import { useQuery } from '@tanstack/react-query';
 import ms from 'ms';
 import genres from '../data/genres';
 import APIClient from '../services/api-client';
+import { Genre } from '../interfaces/Genre';
 
 const apiClient = new APIClient<Genre>('/genres');
-
-export interface Genre {
-  id: number;
-  name: string;
-  image_background: string;
-}
 
 const useGenres = () =>
   useQuery({
     queryKey: ['genres'],
     queryFn: apiClient.getAll,
-    staleTime: ms('1d') , // 24 hours
+    staleTime: ms('1d'), // 24 hours
     initialData: genres,
   });
