+++ src/hooks/usePlatform.ts
@@ -2,21 +2,16 @@ import { useQuery } from '@tanstack/react-query';
 import ms from 'ms';
 import platforms from '../data/platforms';
 import APIClient from '../services/api-client';
+import { Platform } from '../interfaces/Platform';
 
 const apiClient = new APIClient<Platform>('/platforms/list/parents');
-
-export interface Platform {
-  id: number;
-  name: string;
-  slug: string;
-}
 
 const usePlatform = () =>
   useQuery({
     queryKey: ['platforms'],
     queryFn: apiClient.getAll,
     staleTime: ms('1d'),
-    initialData: platforms
+    initialData: platforms,
   });
 
 export default usePlatform;
