+++ src/interfaces/Game.ts
@@ -1,0 +1,12 @@
+import { Platform } from './Platform';
+
+export interface Game {
+  id: number;
+  slug: string;
+  name: string;
+  description_raw: string;
+  background_image: string;
+  parent_platforms: { platform: Platform }[];
+  metacritic: number;
+  rating_top: number;
+}
