+++ src/interfaces/Platform.ts
@@ -1,0 +1,5 @@
+export interface Platform {
+  id: number;
+  name: string;
+  slug: string;
+}
